Moved speech recognition modules to new directory structure
